Basic functionality implemented
Need to make it so that only one stack of an item is stabilized. This can wait until after i've implemented the "stabilized" itemtype.
Also doesn't seem to work with vanilla void, for some reason? Perhaps this mod loaded before vanilla void. Will have to add VV as a dependency to see if that fixes the issue.